some refactoring, begin tree traversal
Commented out the alternative related attributes from the Criteria
class in anticipation of factoring them out of it entirely in the
future.

Created general.py, which contains a function for tree traversal.
Did not add to the lib file b/c it's generally useful.  Will make
general.py into a general "tools" library in future.